--- JSONParser/JSONParser/Utility/ErrorCode.swift
@@ -1,14 +1,0 @@
-//
-//  ErrorCode.swift
-//  JSONParser
-//
-//  Created by Mrlee on 2017. 11. 13..
-//  Copyright © 2017년 JK. All rights reserved.
-//
-
-import Foundation
-
-enum ErrorCode: Error {
-    case invalidInputString
-    case invalidJSONStandard
-}
